jit: avoid wraparound in stack size
pcre2_jit_stack_create() allows the user to indicate how big of a
stack size, JIT should be able to allocate and use, using a size_t
variable that should be able to hold bigger values than reasonable.

Internally, the value is rounded to the next 8K, but if the value
is unreasonable large, would overflow.

Avoid the overflowing by checking and failing the value early, and
while at it make the check clearer and document the failure mode.

Signed-off-by: Carlo Marcelo Arenas Belón <carenas@gmail.com>